requirements.txt
asgiref==3.3.1
coverage==5.3.1
Django==3.1.3
django-debug-toolbar==3.2
Pillow==8.0.1
python-slugify==4.0.1
pytz==2020.4
sqlparse==0.4.1
text-unidecode==1.3
view==0.1
